device delete and new

+++ server/admin/devices.go
@@ -69,6 +69,9 @@ func NewDevice(c *gin.Context) {
 		internal.ErrReqPara(c, err)
 		return
 	}
+	if device.Id == "" {
+		device.Id = uuid.NewString()
+	}
 	device.TenantId = tenant.Id
 	if err := global.DB.Create(&device).Error; err != nil {
 		c.Status(http.StatusConflict)
@@ -77,6 +80,7 @@ func NewDevice(c *gin.Context) {
 	}
 
 	secret := models.DeviceSecret{
+		DeviceId: device.Id,
 		Name:     "default",
 		Secret:   uuid.NewString(),
 		TenantId: tenant.Id,
@@ -144,6 +148,13 @@ func DeleteDevice(c *gin.Context) {
 		global.LOG.Error("get device err: " + err.Error())
 		return
 	}
+	if err := global.DB.Where("device_id = ?", deviceId).
+		Delete(&models.DeviceSecret{}).Error; err != nil {
+		c.Status(http.StatusInternalServerError)
+		global.LOG.Error("delete device err: " + err.Error())
+		return
+	}
+
 	if err := global.DB.Delete(&device).Error; err != nil {
 		c.Status(http.StatusInternalServerError)
 		global.LOG.Error("delete device err: " + err.Error())
